% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dictionary.R
\name{check_dictionary}
\alias{check_dictionary}
\title{Check data dictionary}
\usage{
check_dictionary(dictionary, min_nchars = 2)
}
\arguments{
\item{dictionary}{a list tables corresponding to individual data dictionary tables}

\item{min_nchars}{the minimum number length of characters required (default = 2)
for the display_name, display_name_label and field_description
fields}
}
\value{
list comprising of cleaned data dictionary tables and information about
various column level checks
}
\description{
This function completes a number of checks on the dictionary and highlights
rows which invalidate any checks.
}
