% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dictionary.R
\name{check_for_expected_vals}
\alias{check_for_expected_vals}
\title{Check a field against a vector of specified values}
\usage{
check_for_expected_vals(dictionary, field, expected_vals)
}
\arguments{
\item{dictionary}{a data.frame corresponding to an individual data dictionary table}

\item{field}{the name of the field to search}

\item{expected_vals}{a character vector of values to search against}
}
\value{
logical
}
\description{
Check a field against a vector of specified values
}
