% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dictionary.R
\name{check_min_nchar}
\alias{check_min_nchar}
\title{Check if a field contains a minimum chars}
\usage{
check_min_nchar(dictionary, field, nchars = 2)
}
\arguments{
\item{dictionary}{a data.frame corresponding to an individual data dictionary table}

\item{field}{the name of the field to search}

\item{nchars}{the minimum number length of characters required (default = 2)}
}
\value{
logical
}
\description{
Check if a field contains a minimum chars
}
