% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_fields.R
\name{check_missing_fields}
\alias{check_missing_fields}
\title{Check missing fields}
\usage{
check_missing_fields(dictionary, reference)
}
\arguments{
\item{dictionary}{a list of data.frames corresponding to an individual data dictionary tables}

\item{reference}{a list of character vectors corresponding to field names
which will be compared with those in each dictionary table}
}
\value{
either a logical when no missing fields were detected or a list characters
corresponding to fields missing per each specific table
}
\description{
Compare the dictionary fields against a reference to identify fields which are
missing in the dictionary
}
