% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dictionary.R
\name{check_value_lengths}
\alias{check_value_lengths}
\title{Check if all values are the same length}
\usage{
check_value_lengths(dictionary, field)
}
\arguments{
\item{dictionary}{a data.frame corresponding to an individual data dictionary table}

\item{field}{the name of the field containing the values}
}
\value{
logical
}
\description{
Check if all values are the same length
}
