% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_customer.R
\name{clean_environment_fields}
\alias{clean_environment_fields}
\title{Clean environment fields}
\usage{
clean_environment_fields(fields)
}
\arguments{
\item{fields}{a data.frame of fields present in a customer's agreement}
}
\value{
list of fields present in a customers agreement split by table name
}
\description{
This function cleans the table of fields present in a customer's agreement
}
