% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dictionary_tables.R
\name{clean_table}
\alias{clean_table}
\title{Clean an individual dictionary table}
\usage{
clean_table(dd_table, col_order, agreement, table_name)
}
\arguments{
\item{dd_table}{a data.frame corresponding to a specific dictionary dataset table}

\item{col_order}{a character vector detailing the order of the columns}

\item{agreement}{a character representing the customer's agreement id}

\item{table_name}{a character representing the specific table name}
}
\value{
a cleaned data.frame
}
\description{
This function applies a number of formatting options to an individual
dictionary table. These include:
1.) Sorting the column order;
2.) Converting variable and data types to title case;
3.) Adding dataset/agreement specific path and table names;
4.) Adding a new line return after each "value" key-value pair.
}
