% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table_names.R
\name{clean_table_names}
\alias{clean_table_names}
\title{Clean data dictionary table names}
\usage{
clean_table_names(table_names, remove_fyear = FALSE)
}
\arguments{
\item{table_names}{a character vector of table names}

\item{remove_fyear}{logical indicating if the financial year components should be removed}
}
\value{
character vector of cleaned table names
}
\description{
This function cleans and standardises the data dictionary table names to
include/remove the financial year components and also accounts for the maximum
excel tab sheet name length (n = 31)
}
