% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_workbook.R
\name{create_formatted_workbook}
\alias{create_formatted_workbook}
\title{Create a formatted excel workbook}
\usage{
create_formatted_workbook(
  dictionary,
  hometab_filepath,
  customer_name,
  customer_agreement,
  outfilepath,
  overwrite_existing = TRUE
)
}
\arguments{
\item{dictionary}{a list of cleaned data dictionary tables}

\item{hometab_filepath}{a file path where the home tab excel workbook is located}

\item{customer_name}{a character corresponding to the customers name}

\item{customer_agreement}{a character corresponding to the customers agreement id}

\item{outfilepath}{a file path where you would like to save the workbook. The "/" should be omitted in this path}

\item{overwrite_existing}{logical should an existing file be overwritten (default: TRUE)}
}
\value{
a formatted excel workbook saved to the outfilepath
}
\description{
This function will take a list cleaned data dictionary tables and return
a formatted excel workbook. It will also allows functionality to add a home tab
to the final document.
}
