% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tables.R
\name{create_split_table}
\alias{create_split_table}
\title{Split a data dictionary table based on the fields within a customer table}
\usage{
create_split_table(
  dictionary_tables,
  split_lookups,
  split_field = "display_name"
)
}
\arguments{
\item{dictionary_tables}{a list of data.frames where each element corresponds
to a specific data dictionary table}

\item{split_lookups}{list of split lookup values generated by \code{get_all_split_keys}
which will be split}

\item{split_field}{the name of the field that will be used to create the split tables}
}
\value{
a list of split data.frames
}
\description{
For a given data dictionary table this function will split the table
based the common values found within a user-defined field
}
