% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_split_table_keys.R
\name{.check_for_missing_keys}
\alias{.check_for_missing_keys}
\title{check for missing keys between a customer table and a dictionary table}
\usage{
.check_for_missing_keys(split_lookup)
}
\arguments{
\item{split_lookup}{a list of lookup keys generated by \code{get_split_keys}}
}
\description{
check for missing keys between a customer table and a dictionary table
}
