% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dictionary_tables.R
\name{fields_present}
\alias{fields_present}
\title{Test for expected fields}
\usage{
fields_present(dictionary, expected)
}
\arguments{
\item{dictionary}{a list of dictionary tables}

\item{expected}{a character vector of expected field names}
}
\value{
logical if all fields are present
}
\description{
This function tests if all expected fields are present in each of the the
dictionary table
}
