% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_split_tables.R
\name{find_tables_to_split}
\alias{find_tables_to_split}
\title{Search a list of customer tables and identify if any tables are from the cancer dataset}
\usage{
find_tables_to_split(customer_tables, dictionary_tables)
}
\arguments{
\item{customer_tables}{a list of data.frames where each element corresponds to a
specific customer table, based on their field list information}

\item{dictionary_tables}{a list of data.frames where each element corresponds
to a specific data dictionary table}
}
\value{
\item{NULL}{If no cancer datasets were found}

OR

List where each element corresponds to a cancer datasets table and has the following elements:

\item{customer}{character vector of customer tables (length >=1)}
\item{dd}{A character vector of the data dictionary table (length == 1)}
}
\description{
Search a list of customer tables and identify if any tables are from the cancer dataset
}
\details{
The cancer dataset currently, comprises of the following tables:
\itemize{
\item National Cancer Registrations (NCR)
\item Rapid Cancer Registration (RCR)
\item Systemic Anti-Cancer Therapy (SACT)
\item Rapid Diagnostic Centre (RDC)
\item Radiotherapy Dataset (RTDS)
}
}
