% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dd.R
\name{generate_dd}
\alias{generate_dd}
\title{Generate a formatted data dictionary}
\usage{
generate_dd(unformatted_dict_path, customer_fields_path, customer_name)
}
\arguments{
\item{unformatted_dict_path}{a file path to an excel workbook containing one unformatted dictionary table per sheet}

\item{customer_fields_path}{a file path corresponding to a csv file which lists
all the fields present in the selected customers agreement}

\item{customer_name}{a character vector corresponding to one of the approved users}
}
\value{
A list of length 4: unformatted_dd = a list of the unformatted dictionary tables before cleaning;
customer_info = a list of 3 comprising customer information and agreement specific fields;
cleaned_dd = a list of cleaned cleaned dictionary tables;checks = a list of checks
completed against each of the cleaned dictionary tables.
}
\description{
Generate a formatted data dictionary
}
