% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dictionary.R
\name{get_all_unique_units}
\alias{get_all_unique_units}
\title{Return all unique units}
\usage{
get_all_unique_units(dictionary, field)
}
\arguments{
\item{dictionary}{a data.frame corresponding to an individual data dictionary table}

\item{field}{the name of the field containing the units}
}
\value{
character of unique unit values
}
\description{
Return all unique units
}
