% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_fields.R
\name{get_common_fields}
\alias{get_common_fields}
\title{Get common fields}
\usage{
get_common_fields(dictionary, reference, field)
}
\arguments{
\item{dictionary}{a list of data.frames corresponding to individual data dictionary tables}

\item{reference}{a list of character vectors corresponding to field names
which will be compared with those in each dictionary table}
}
\value{
A list of subset data.frames
}
\description{
Compare two lists of data.frames and subset based on the common values for
a given field which is present in both data.frames.
}
