% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_split_table_keys.R
\name{get_split_keys}
\alias{get_split_keys}
\title{Retrieve the customer table keys which can be used to split table on.}
\usage{
get_split_keys(
  tables_to_split,
  customer_tables,
  dictionary_tables,
  split_key_field = "display_name"
)
}
\arguments{
\item{tables_to_split}{List of customer and corresponding dictionary tables to
split returned from \code{find_cancer_tables}}

\item{customer_tables}{a list of data.frames where each element corresponds to a
specific customer table, based on their field list information}

\item{dictionary_tables}{a list of data.frames where each element corresponds
to a specific data dictionary table}

\item{split_key_field}{the field within a customer table from which the keys will be retrieved}
}
\value{
List with the same dimensions as the \code{tables_to_split} input with
keys replacing the customer table names.
}
\description{
Retrieve the customer table keys which can be used to split table on.
}
