% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_fields.R
\name{identify_missing_tbl_fields}
\alias{identify_missing_tbl_fields}
\title{Compare fields values against a reference}
\usage{
identify_missing_tbl_fields(
  dictionary_tbl,
  reference_tbl,
  field,
  return_common = FALSE
)
}
\arguments{
\item{dictionary_tbl}{a data.frame corresponding containing a field which you want to compare against a reference}

\item{reference_tbl}{a data.frame containing a reference field which you wish to compare}

\item{field}{the name of the field to compare in both the reference and dictionary table}

\item{return_common}{if TRUE (default: FALSE) returns the common fields between the dictionary and the reference}
}
\value{
logical or a list of characters corresponding to field names (either the difference or intersect)
}
\description{
Compare a fields values against a set of reference values to identify which are
missing or in common between the two.
}
