% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_split_table_keys.R
\name{missing_split_keys}
\alias{missing_split_keys}
\title{Identify missing split keys between customer and dictionary tables}
\usage{
missing_split_keys(split_lookup_keys)
}
\arguments{
\item{split_lookup_keys}{a list of lookup keys generated by \code{get_split_keys}}
}
\value{
List of length \code{split_lookup_keys} a logical of length 1, equalling
FALSE to denote that there are no missing keys in the split keys lookup

OR

a list of character vectors, one for each set lookup tables, which displays any
keys which are missing between the customer table and the corresponding
dictionary table
}
\description{
Identify missing split keys between customer and dictionary tables
}
