% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_fields.R
\name{missing_tbl_fields}
\alias{missing_tbl_fields}
\title{Compare fields with against a reference}
\usage{
missing_tbl_fields(dictionary_tbl, reference_tbl, field)
}
\arguments{
\item{dictionary_tbl}{a data.frame corresponding to an individual data dictionary table}

\item{reference_tbl}{a data.frame containing a column of reference values to compare against}

\item{field}{the name of the field to compare in both the reference and dictionary}
}
\value{
logical
}
\description{
Compare fields against a reference to identify  fields which are
missing in the dictionary table
}
