% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_sheets.R
\name{read_excel_sheets}
\alias{read_excel_sheets}
\title{Read in an excel workbook}
\usage{
read_excel_sheets(workbook_path, filter_sheets)
}
\arguments{
\item{workbook_path}{path to the excel workbook}

\item{filter_sheets}{character vector of sheet names to filter out}
}
\value{
A list of data frames corresponding to individual excel workbook sheets
}
\description{
This function reads in an excel workbook into R.
}
