% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_missing_tables.R
\name{remove_missing_tables}
\alias{remove_missing_tables}
\title{Remove missing table}
\usage{
remove_missing_tables(dictionary, fields)
}
\arguments{
\item{dictionary}{a list of data dictionary tables returned by \code{\link[=clean_table_names]{clean_table_names()}}}

\item{fields}{a list of fields present in a users agreement, returned from \code{\link[=select_customer]{select_customer()}}}
}
\value{
list of dictionary and field tables
}
\description{
This function removes any tables (from the data dictionary) that are not
present within the list of fields for a given customers data sharing agreement.
}
