% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREDD-package.R
\docType{package}
\name{TREDD-package}
\alias{TREDD}
\alias{TREDD-package}
\title{TREDD: Trusted Research Environment Service Data Dictionary}
\description{
This package contains various functions which can be used to generate a formatted data dictionary and check for discrepancies in the unformatted dictionaries.
}
\author{
\strong{Maintainer}: Shoaib Ajaib \email{shoaib.ajaib1@nhs.com}

}
\keyword{internal}
