% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dictionary_tables.R
\name{clean_dictionary_tables}
\alias{clean_dictionary_tables}
\title{Clean data dictionary}
\usage{
clean_dictionary_tables(
  dictionary,
  agreement_id,
  expected_fields = expected$fields
)
}
\arguments{
\item{dictionary}{a list tables corresponding to individual data dictionary tables}

\item{agreement_id}{a character representing the customer's agreement id}

\item{expected_fields}{a character vector of expected field names}
}
\value{
list comprising of cleaned data dictionary tables and information about
various column level checks
}
\description{
This function cleans the unformatted list of data dictionary tables and also
tests various table columns to ensure they meet each a pre-defined criteria.
}
