% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_and_combine_tables.R
\name{combine_split_tables}
\alias{combine_split_tables}
\title{combine dictionary tables into one table}
\usage{
combine_split_tables(dictionary, dataset)
}
\arguments{
\item{dictionary}{a list of data.frames containing the tables to be combined}

\item{dataset}{a character vector of the dataset to be split (currently this is only the mental health dataset)}
}
\value{
a list of data.frames or NULL
}
\description{
combine dictionary tables into one table
}
