% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_split_tables.R
\name{find_tables_to_split}
\alias{find_tables_to_split}
\title{Search a list of customer tables using a split table lookup list and identify
if any tables that require splitting into multiple tables are present. The
default split table lookup list is defined in \code{split_tables_lookup}, but can
be editing and overwritten by the user and then passed to the function.}
\usage{
find_tables_to_split(
  customer_tables,
  dictionary_tables,
  split_tables_lookup = split_tables_lookup
)
}
\arguments{
\item{customer_tables}{a list of data.frames where each element corresponds to a
specific table that is provisioned for a customer. This table
will also comprise a list of fields that are present within
the table.}

\item{dictionary_tables}{a list of data.frames where each element corresponds
to a specific data dictionary table found within the
unformatted data dictionary excel file.}

\item{split_tables_lookup}{a list of character vectors (length = 2) where each
element corresponds to a table that is to be
split into multiple tables. the first element
contains a regular expression that is used to identify
tables that are to be split. The second element contains
a regular expression that identifies the data dictionary
table that is to be split.}
}
\value{
\item{NULL}{If no cancer datasets were found}

OR

List where each element corresponds to a cancer datasets table and has the following elements:

\item{customer}{character vector of customer tables (length >=1)}
\item{dd}{A character vector of the data dictionary table (length == 1)}
}
\description{
Search a list of customer tables using a split table lookup list and identify
if any tables that require splitting into multiple tables are present. The
default split table lookup list is defined in \code{split_tables_lookup}, but can
be editing and overwritten by the user and then passed to the function.
}
\details{
The cancer dataset currently, comprises of the following tables:
\itemize{
\item National Cancer Registrations (NCR)
\item Rapid Cancer Registration (RCR)
\item Systemic Anti-Cancer Therapy (SACT)
\item Rapid Diagnostic Centre (RDC)
\item Radiotherapy Dataset (RTDS)
}
}
